♻️ REFACTOR: Remove `QueryManager`
