+++ aiida/cmdline/commands/cmd_data/cmd_bands.py
@@ -41,11 +41,10 @@ def bands():
 @options.FORMULA_MODE()
 def bands_list(elements, elements_exclusive, raw, formula_mode, past_days, groups, all_users):
     """List BandsData objects."""
-    from aiida.manage.manager import get_manager
     from tabulate import tabulate
     from argparse import Namespace
 
-    backend = get_manager().get_backend()
+    from aiida.orm.nodes.data.array.bands import get_bands_and_parents_structure
 
     args = Namespace()
     args.element = elements
@@ -59,8 +58,7 @@ def bands_list(elements, elements_exclusive, raw, formula_mode, past_days, group
         args.group_pk = None
     args.all_users = all_users
 
-    query = backend.query_manager
-    entry_list = query.get_bands_and_parents_structure(args)
+    entry_list = get_bands_and_parents_structure(args)
 
     counter = 0
     bands_list_data = list()
