+++ aiida/orm/nodes/data/array/bands.py
@@ -1799,3 +1799,135 @@ MATPLOTLIB_FOOTER_TEMPLATE_SHOW = Template("""pl.show()""")
 MATPLOTLIB_FOOTER_TEMPLATE_EXPORTFILE = Template("""pl.savefig("$fname", format="$format")""")
 
 MATPLOTLIB_FOOTER_TEMPLATE_EXPORTFILE_WITH_DPI = Template("""pl.savefig("$fname", format="$format", dpi=$dpi)""")
+
+
+def get_bands_and_parents_structure(args):
+    """Search for bands and return bands and the closest structure that is a parent of the instance.
+
+    :returns:
+        A list of sublists, each latter containing (in order):
+            pk as string, formula as string, creation date, bandsdata-label
+    """
+    # pylint: disable=too-many-locals
+
+    import datetime
+    from aiida.common import timezone
+    from aiida import orm
+
+    q_build = orm.QueryBuilder()
+    if args.all_users is False:
+        q_build.append(orm.User, tag='creator', filters={'email': orm.User.objects.get_default().email})
+    else:
+        q_build.append(orm.User, tag='creator')
+
+    group_filters = {}
+
+    if args.group_name is not None:
+        group_filters.update({'name': {'in': args.group_name}})
+    if args.group_pk is not None:
+        group_filters.update({'id': {'in': args.group_pk}})
+
+    q_build.append(orm.Group, tag='group', filters=group_filters, with_user='creator')
+
+    bdata_filters = {}
+    if args.past_days is not None:
+        bdata_filters.update({'ctime': {'>=': timezone.now() - datetime.timedelta(days=args.past_days)}})
+
+    q_build.append(
+        orm.BandsData, tag='bdata', with_group='group', filters=bdata_filters, project=['id', 'label', 'ctime']
+    )
+    bands_list_data = q_build.all()
+
+    q_build.append(
+        orm.StructureData,
+        tag='sdata',
+        with_descendants='bdata',
+        # We don't care about the creator of StructureData
+        project=['id', 'attributes.kinds', 'attributes.sites']
+    )
+
+    q_build.order_by({orm.StructureData: {'ctime': 'desc'}})
+
+    structure_dict = dict()
+    list_data = q_build.distinct().all()
+    for bid, _, _, _, akinds, asites in list_data:
+        structure_dict[bid] = (akinds, asites)
+
+    entry_list = []
+    already_visited_bdata = set()
+
+    for [bid, blabel, bdate] in bands_list_data:
+
+        # We process only one StructureData per BandsData.
+        # We want to process the closest StructureData to
+        # every BandsData.
+        # We hope that the StructureData with the latest
+        # creation time is the closest one.
+        # This will be updated when the QueryBuilder supports
+        # order_by by the distance of two nodes.
+        if already_visited_bdata.__contains__(bid):
+            continue
+        already_visited_bdata.add(bid)
+        strct = structure_dict.get(bid, None)
+
+        if strct is not None:
+            akinds, asites = strct
+            formula = _extract_formula(akinds, asites, args)
+        else:
+            if args.element is not None or args.element_only is not None:
+                formula = None
+            else:
+                formula = '<<NOT FOUND>>'
+
+        if formula is None:
+            continue
+        entry_list.append([str(bid), str(formula), bdate.strftime('%d %b %Y'), blabel])
+
+    return entry_list
+
+
+def _extract_formula(akinds, asites, args):
+    """
+    Extract formula from the structure object.
+
+    :param akinds: list of kinds, e.g. [{'mass': 55.845, 'name': 'Fe', 'symbols': ['Fe'], 'weights': [1.0]},
+                                        {'mass': 15.9994, 'name': 'O', 'symbols': ['O'], 'weights': [1.0]}]
+    :param asites: list of structure sites e.g. [{'position': [0.0, 0.0, 0.0], 'kind_name': 'Fe'},
+                                                    {'position': [2.0, 2.0, 2.0], 'kind_name': 'O'}]
+    :param args: a namespace with parsed command line parameters, here only 'element' and 'element_only' are used
+    :type args: dict
+
+    :return: a string with formula if the formula is found
+    """
+    from aiida.orm.nodes.data.structure import (get_formula, get_symbols_string)
+
+    if args.element is not None:
+        all_symbols = [_['symbols'][0] for _ in akinds]
+        if not any(s in args.element for s in all_symbols):
+            return None
+
+    if args.element_only is not None:
+        all_symbols = [_['symbols'][0] for _ in akinds]
+        if not all(s in all_symbols for s in args.element_only):
+            return None
+
+    # We want only the StructureData that have attributes
+    if akinds is None or asites is None:
+        return '<<UNKNOWN>>'
+
+    symbol_dict = {}
+    for k in akinds:
+        symbols = k['symbols']
+        weights = k['weights']
+        symbol_dict[k['name']] = get_symbols_string(symbols, weights)
+
+    try:
+        symbol_list = []
+        for site in asites:
+            symbol_list.append(symbol_dict[site['kind_name']])
+        formula = get_formula(symbol_list, mode=args.formulamode)
+    # If for some reason there is no kind with the name
+    # referenced by the site
+    except KeyError:
+        formula = '<<UNKNOWN>>'
+    return formula
